+++ previous-versions/files/src/codatfiles/models/operations/__init__.py
@@ -4,4 +4,4 @@ from .download_files import *
 from .list_files import *
 from .upload_files import *
 
-__all__ = ["DownloadFilesErrorMessage","DownloadFilesRequest","DownloadFilesResponse","ListFilesErrorMessage","ListFilesRequest","ListFilesResponse","UploadFilesErrorMessage","UploadFilesRequest","UploadFilesRequestBody","UploadFilesResponse"]
+__all__ = ["DownloadFilesRequest","DownloadFilesResponse","ListFilesRequest","ListFilesResponse","UploadFilesRequest","UploadFilesRequestBody","UploadFilesResponse"]