+++ previous-versions/files/RELEASES.md
@@ -9,3 +9,13 @@ Based on:
 - [python v0.1.0] previous-versions/files
 ### Releases
 - [PyPI v0.1.0] https://pypi.org/project/codat-files/0.1.0 - previous-versions/files
+
+## 2023-12-11 01:27:03
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Files.yaml
+- Speakeasy CLI 1.125.2 (2.210.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.2.0] previous-versions/files
+### Releases
+- [PyPI v0.2.0] https://pypi.org/project/codat-files/0.2.0 - previous-versions/files